agents/AI-Researcher-Analyzer-Agent/requirements.txt
# requirements.txt
langchain
langchain_community
langgraph
python-dotenv
openai
linkedin-api  # For LinkedIn access
loguru
pydantic
pytest
python-decouple
sqlalchemy
fastapi
uvicorn
requests
langchain[groq]
selenium 
beautifulsoup4
pandas
playwright
redis
supabase
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
python-dotenv
mysql-connector-python
requests
fastapi
uvicorn
python-jose[cryptography]
authlib
pyjwt
itsdangerous
google-auth
selenium-stealth
undetected-chromedriver